Move ManageSpace* to .ui.windows.managespace
As a part of #12582 initiative
